Move more files into better locations
